finish passing the tests; remove ProtocolInterface
